+++ src/Application/Interfaces/IAuthService.cs
@@ -4,7 +4,7 @@ namespace TechCurse.Application.Interfaces;
 
 public interface IAuthService
 {
-    Task<bool> RegisterAsync(RegisterInputDto input);
+    Task RegisterAsync(RegisterInputDto input);
     Task CreateUserAsync(CreateUserInputDto input);
     Task<AuthOutputDto?> LoginAsync(LoginInputDto input);
     Task<AuthOutputDto?> RefreshAsync(RefreshTokenInputDto input);